E2e refactoring: retry mechanisms
